Debugging 4/13 (Morning)
Isolated Error Array Error: Caused by the Splitter creating a blank
string before the '$'.  Resolved by placeing the splitter on the other
side of the element.

Currently working on fixing the Trim function as it is not removing the
'>' from the end of the Saved GPS Data.  This causes the program to
conitnually add '>' to the end each time it runs the
GPSlocation.ToString() method as it conciders '>' part of the value of
the last property in CusomInfo